@@ -12,11 +12,12 @@
 * limitations under the License.
 */
 import { HttpClientModule } from "@angular/common/http";
-import { ComponentFixture, TestBed } from "@angular/core/testing";
+import { type ComponentFixture, TestBed, fakeAsync } from "@angular/core/testing";
 import { ReactiveFormsModule } from "@angular/forms";
 import { RouterTestingModule } from "@angular/router/testing";
 
-import {CacheGroupService} from "../../../shared/api";
+import { APITestingModule } from "src/app/api/testing";
+
 import { CacheGroupTableComponent } from "./cache-group-table.component";
 
 describe("CacheGroupTableComponent", () => {
@@ -24,18 +25,15 @@ describe("CacheGroupTableComponent", () => {
 	let fixture: ComponentFixture<CacheGroupTableComponent>;
 
 	beforeEach(async () => {
-		const mockAPIService = jasmine.createSpyObj(["getCacheGroups"]);
 		await TestBed.configureTestingModule({
 			declarations: [ CacheGroupTableComponent ],
-			imports: [ReactiveFormsModule, HttpClientModule, RouterTestingModule],
-			providers: [
-				{ provide: CacheGroupService, useValue: mockAPIService }
-			]
-		})
-			.compileComponents();
-	});
-
-	beforeEach(() => {
+			imports: [
+				APITestingModule,
+				HttpClientModule,
+				ReactiveFormsModule,
+				RouterTestingModule
+			],
+		}).compileComponents();
 		fixture = TestBed.createComponent(CacheGroupTableComponent);
 		component = fixture.componentInstance;
 		fixture.detectChanges();
@@ -43,5 +41,15 @@ describe("CacheGroupTableComponent", () => {
 
 	it("should create", () => {
 		expect(component).toBeTruthy();
+	});
+
+	it("emits the search box value", fakeAsync(() => {
+		component.fuzzControl.setValue("query");
+		component.updateURL();
+		expectAsync(component.fuzzySubject.toPromise()).toBeResolvedTo("query");
+	}));
+
+	it("doesn't throw errors when handling context menu events", () => {
+		expect(()=>component.handleContextMenu({action: "something", data: []})).not.toThrow();
 	});
 });
