@@ -185,9 +185,10 @@ function run(): number {
 				key,
 				rejectUnauthorized: !config.insecure
 			},
-			server).listen(config.port, ()=> {
-				console.log(`Node Express server listening on port ${config.port}`);
-			});
+			server
+		).listen(config.port, ()=> {
+			console.log(`Node Express server listening on port ${config.port}`);
+		});
 		try {
 			createRedirectServer(
 				(req, res) => {
