@@ -14,8 +14,9 @@
 import { HttpClient } from "@angular/common/http";
 import { Injectable } from "@angular/core";
 
-import { Type } from "../../models";
-import { APIService } from "./APIService";
+import type { Type } from "src/app/models";
+
+import { APIService } from "./base-api.service";
 
 /** The allowed values for the 'useInTables' query parameter of GET requests to /types. */
 type UseInTable = "cachegroup" |
