@@ -13,38 +13,54 @@
 */
 
 import { HttpClientModule } from "@angular/common/http";
-import { TestBed, waitForAsync } from "@angular/core/testing";
-import { RouterTestingModule } from "@angular/router/testing";
+import { fakeAsync, TestBed, tick } from "@angular/core/testing";
 import { MatSnackBarModule } from "@angular/material/snack-bar";
+import { Router } from "@angular/router";
+import { RouterTestingModule } from "@angular/router/testing";
+import { of } from "rxjs";
 
 import { CurrentUserService } from "src/app/shared/currentUser/current-user.service";
+
 import { AppComponent } from "./app.component";
 
 describe("AppComponent", () => {
-	beforeEach(waitForAsync(() => {
-		const mockCurrentUserService = jasmine.createSpyObj(["updateCurrentUser", "login", "logout"]);
+	let component: AppComponent;
+	let mockCurrentUserService: jasmine.SpyObj<CurrentUserService>;
+
+	beforeEach(() => {
+		mockCurrentUserService = jasmine.createSpyObj(["updateCurrentUser", "login", "logout"], {userChanged: of(null)});
 		TestBed.configureTestingModule({
 			declarations: [
 				AppComponent
 			],
 			imports: [
 				HttpClientModule,
-				RouterTestingModule,
+				RouterTestingModule.withRoutes([{component: AppComponent, path: "login"}]),
 				MatSnackBarModule
 			],
 			providers: [ { provide: CurrentUserService, useValue: mockCurrentUserService }]
 		}).compileComponents();
-	}));
+		const fixture = TestBed.createComponent(AppComponent);
+		component = fixture.componentInstance;
+		fixture.detectChanges();
+	});
 
 	it("should create the app", () => {
-		const fixture = TestBed.createComponent(AppComponent);
-		const app = fixture.componentInstance;
-		expect(app).toBeTruthy();
+		expect(component).toBeTruthy();
 	});
 
 	it("should have as title 'Traffic Portal'", () => {
-		const fixture = TestBed.createComponent(AppComponent);
-		const app = fixture.componentInstance;
-		expect(app.title).toEqual("Traffic Portal");
+		expect(component.title).toEqual("Traffic Portal");
 	});
+
+	it("logs out", fakeAsync(() => {
+		// eslint-disable-next-line @typescript-eslint/unbound-method
+		expect(mockCurrentUserService.logout).not.toHaveBeenCalled();
+		component.logout();
+		tick();
+		// eslint-disable-next-line @typescript-eslint/unbound-method
+		expect(mockCurrentUserService.logout).toHaveBeenCalled();
+		const router = TestBed.inject(Router);
+		expect(router.url).toBe("/login");
+	}));
 });
