@@ -13,10 +13,10 @@
 */
 
 import { Component, OnInit } from "@angular/core";
-import {Router} from "@angular/router";
+import { Router } from "@angular/router";
 
 import { CurrentUser } from "src/app/models";
-import {CurrentUserService} from "src/app/shared/currentUser/current-user.service";
+import { CurrentUserService } from "src/app/shared/currentUser/current-user.service";
 
 /**
  * The most basic component that contains everything else. This should be kept pretty simple.
@@ -33,9 +33,6 @@ export class AppComponent implements OnInit {
 	/** The currently logged-in user */
 	public currentUser: CurrentUser | null = null;
 
-	/**
-	 * Constructor.
-	 */
 	constructor(private readonly router: Router, private readonly auth: CurrentUserService) {
 	}
 
