@@ -11,14 +11,15 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
-import { Component, OnInit } from "@angular/core";
+import { Component, type OnInit } from "@angular/core";
 import { MatDialog } from "@angular/material/dialog";
 import { ActivatedRoute, Router } from "@angular/router";
 import { faEdit } from "@fortawesome/free-solid-svg-icons";
-import { UserService } from "src/app/shared/api";
 
-import { CurrentUser } from "src/app/models";
-import {CurrentUserService} from "src/app/shared/currentUser/current-user.service";
+import { UserService } from "src/app/api";
+import type { CurrentUser } from "src/app/models";
+import { CurrentUserService } from "src/app/shared/currentUser/current-user.service";
+
 import { UpdatePasswordDialogComponent } from "./update-password-dialog/update-password-dialog.component";
 
 /**
@@ -87,8 +88,7 @@ export class CurrentuserComponent implements OnInit {
 	 */
 	public edit(): void {
 		if (!this.currentUser) {
-			console.error("cannot edit null user");
-			return;
+			throw new Error("cannot edit null user");
 		}
 		this.editUser = {...this.currentUser};
 		this.editing = true;
@@ -125,7 +125,7 @@ export class CurrentuserComponent implements OnInit {
 	 *
 	 * @param e The form submittal event.
 	 */
-	public submitEdit(e: Event): void {
+	public async submitEdit(e: Event): Promise<void> {
 		e.preventDefault();
 		e.stopPropagation();
 		if (this.editUser === null) {
@@ -136,23 +136,17 @@ export class CurrentuserComponent implements OnInit {
 		this.editUser.localPasswd = undefined;
 		this.editUser.confirmLocalPasswd = undefined;
 
-		this.api.updateCurrentUser(this.editUser).then(
-			success => {
-				if (success) {
-					this.auth.updateCurrentUser().then(
-						updated => {
-							if (!updated) {
-								console.warn("Failed to fetch current user after successful update");
-							}
-							this.currentUser = this.auth.currentUser;
-							this.cancelEdit();
-						}
-					);
-				} else {
-					console.warn("Editing the current user failed");
-					this.cancelEdit();
-				}
+		const success = await this.api.updateCurrentUser(this.editUser);
+		if (success) {
+			const updated = await this.auth.updateCurrentUser();
+			if (!updated) {
+				console.warn("Failed to fetch current user after successful update");
 			}
-		);
+			this.currentUser = this.auth.currentUser;
+			this.cancelEdit();
+		} else {
+			console.warn("Editing the current user failed");
+			this.cancelEdit();
+		}
 	}
 }
