@@ -15,7 +15,7 @@ import { Component } from "@angular/core";
 import { MatDialogRef } from "@angular/material/dialog";
 import { Subject } from "rxjs";
 
-import {CurrentUserService} from "src/app/shared/currentUser/current-user.service";
+import { CurrentUserService } from "src/app/shared/currentUser/current-user.service";
 
 /**
  * This is the controller for the "Update Password" dialog box/form.
