@@ -12,83 +12,82 @@
 * limitations under the License.
 */
 import { HttpClientModule } from "@angular/common/http";
-import { waitForAsync, ComponentFixture, TestBed } from "@angular/core/testing";
+import { type ComponentFixture, TestBed } from "@angular/core/testing";
 import { FormsModule, ReactiveFormsModule } from "@angular/forms";
+import { MatDialog, MatDialogModule } from "@angular/material/dialog";
+import { Router } from "@angular/router";
 import { RouterTestingModule } from "@angular/router/testing";
-import { MatDialogModule } from "@angular/material/dialog";
+import { type Observable, of } from "rxjs";
 
-import { of } from "rxjs";
-import {DeliveryServiceService, InvalidationJobService, UserService} from "src/app/shared/api";
-
-
+import { DeliveryServiceService, InvalidationJobService } from "src/app/api";
+import { APITestingModule } from "src/app/api/testing";
+import { defaultDeliveryService, type InvalidationJob, JobType } from "src/app/models";
 import { CurrentUserService } from "src/app/shared/currentUser/current-user.service";
-import { CustomvalidityDirective } from "../../shared/validation/customvalidity.directive";
-import { OpenableDirective } from "../../shared/openable/openable.directive";
-import { DeliveryService, GeoLimit, GeoProvider, InvalidationJob, JobType } from "../../models";
-import {TpHeaderComponent} from "../../shared/tp-header/tp-header.component";
+import { TpHeaderComponent } from "src/app/shared/tp-header/tp-header.component";
+import { CustomvalidityDirective } from "src/app/shared/validation/customvalidity.directive";
+
 import { InvalidationJobsComponent } from "./invalidation-jobs.component";
 
 describe("InvalidationJobsComponent", () => {
 	let component: InvalidationJobsComponent;
 	let fixture: ComponentFixture<InvalidationJobsComponent>;
+	let router: Router;
+	let job: InvalidationJob;
 
-	beforeEach(waitForAsync(() => {
+	beforeEach(async () => {
 		// mock the API
-		const mockAPIService = jasmine.createSpyObj(["getInvalidationJobs", "getDeliveryServices"]);
 		const mockCurrentUserService = jasmine.createSpyObj(["updateCurrentUser", "login", "logout"]);
-		mockAPIService.getInvalidationJobs.and.returnValue(of({
-			startTime: new Date(),
-		} as InvalidationJob));
-		mockAPIService.getDeliveryServices.and.returnValue(of({
-			active: true,
-			anonymousBlockingEnabled: false,
-			cdnId: 0,
-			displayName: "test DS",
-			dscp: 0,
-			geoLimit: GeoLimit.NONE,
-			geoProvider: GeoProvider.MAX_MIND,
-			ipv6RoutingEnabled: true,
-			lastUpdated: new Date(),
-			logsEnabled: true,
-			longDesc: "A test Delivery Service for API mock-ups",
-			missLat: 0,
-			missLong: 0,
-			multiSiteOrigin: false,
-			regionalGeoBlocking: false,
-			routingName: "test-DS",
-			typeId: 0,
-			xmlId: "test-DS"
-		} as DeliveryService));
 
-		TestBed.configureTestingModule({
+		await TestBed.configureTestingModule({
 			declarations: [
 				InvalidationJobsComponent,
 				TpHeaderComponent,
-				OpenableDirective,
 				CustomvalidityDirective
 			],
 			imports: [
+				APITestingModule,
 				FormsModule,
 				HttpClientModule,
 				ReactiveFormsModule,
-				RouterTestingModule,
+				RouterTestingModule.withRoutes([
+					{component: InvalidationJobsComponent, path: "deliveryservice/:id/invalidation-jobs"}
+				]),
 				MatDialogModule
 			],
 			providers: [
-				{ provide: InvalidationJobService, useValue: mockAPIService },
-				{ provide: DeliveryServiceService, useValue: mockAPIService },
-				{ provide: UserService, useValue: mockAPIService },
-				{ provide: CurrentUserService, useValue: mockCurrentUserService }
+				{ provide: CurrentUserService, useValue: mockCurrentUserService },
+				{ provide: MatDialog, useValue: {open: (): {afterClosed: () => Observable<unknown>} => ({
+					afterClosed: () => of(true)
+				})}}
 			]
+		}).compileComponents();
+
+		const dsService = TestBed.inject(DeliveryServiceService);
+		const ds = await dsService.createDeliveryService({...defaultDeliveryService});
+
+		router = TestBed.inject(Router);
+		router.initialNavigation();
+		const navigated = await router.navigate(["/deliveryservice", ds.id, "invalidation-jobs"]);
+		if (!navigated) {
+			return fail("navigation failed");
+		}
+		expect(router.url).toBe(`/deliveryservice/${ds.id}/invalidation-jobs`);
+
+		const jobService = TestBed.inject(InvalidationJobService);
+		job = await jobService.createInvalidationJob({
+			deliveryService: ds.xmlId,
+			regex: "/",
+			startTime: new Date(),
+			ttl: 178
 		});
 
-		TestBed.compileComponents();
-	}));
-
-	beforeEach(() => {
 		fixture = TestBed.createComponent(InvalidationJobsComponent);
 		component = fixture.componentInstance;
 		fixture.detectChanges();
+	});
+
+	afterEach(async ()=> {
+		await expectAsync(component.deleteJob(job.id)).toBeResolved();
 	});
 
 	it("should create", () => {
@@ -110,15 +109,36 @@ describe("InvalidationJobsComponent", () => {
 		j.startTime = new Date(component.now);
 		j.startTime.setMinutes(j.startTime.getMinutes()-30);
 		expect(component.isInProgress(j)).toBeTrue();
-		j.startTime = new Date();
+		j.startTime.setMinutes(j.startTime.getMinutes()+31);
 		expect(component.isInProgress(j)).toBeFalse();
 	});
 
-	afterAll(() => {
-		try{
-			TestBed.resetTestingModule();
-		} catch (e) {
-			console.error("error in InvalidationJobsComponent afterAll:", e);
-		}
+	it("calculates end dates", () => {
+		const j = {
+			assetUrl: "doesn't matter",
+			createdBy: "also doesn't matter",
+			deliveryService: "doesn't matter either",
+			id: -1,
+			keyword: JobType.PURGE,
+			parameters: "",
+			startTime: new Date(0),
+		};
+		expect(()=>component.endDate(j)).toThrow();
+
+		j.parameters = "TTL";
+		expect(()=>component.endDate(j)).toThrow();
+
+		j.parameters = "TTL:not a number";
+		expect(()=>component.endDate(j)).toThrow();
+
+		const expected = new Date(0);
+		expected.setHours(expected.getHours()+178);
+		j.parameters = "TTL:178h";
+		expect(component.endDate(j)).toEqual(expected);
+	});
+
+	it("opens the create/edit dialog", () => {
+		component.editJob(job);
+		component.newJob();
 	});
 });
