@@ -11,13 +11,14 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
-import { Component, OnInit } from "@angular/core";
+import { Component, type OnInit } from "@angular/core";
 import { MatDialog } from "@angular/material/dialog";
 import { ActivatedRoute } from "@angular/router";
 import { faPlus, faTrash, faPencilAlt } from "@fortawesome/free-solid-svg-icons";
 
-import { defaultDeliveryService, DeliveryService, InvalidationJob } from "../../models";
-import { DeliveryServiceService, InvalidationJobService } from "../../shared/api";
+import { DeliveryServiceService, InvalidationJobService } from "src/app/api";
+import { defaultDeliveryService, type DeliveryService, type InvalidationJob } from "src/app/models";
+
 import { NewInvalidationJobDialogComponent } from "./new-invalidation-job-dialog/new-invalidation-job-dialog.component";
 
 /**
