@@ -16,11 +16,12 @@ import { waitForAsync, ComponentFixture, TestBed } from "@angular/core/testing";
 import { FormsModule, ReactiveFormsModule } from "@angular/forms";
 import { RouterTestingModule } from "@angular/router/testing";
 
+import { APITestingModule } from "src/app/api/testing";
+import type { User } from "src/app/models";
 import { CurrentUserService } from "src/app/shared/currentUser/current-user.service";
-import { User } from "../../models";
-import {TpHeaderComponent} from "../../shared/tp-header/tp-header.component";
-import {LoadingComponent} from "../../shared/loading/loading.component";
-import {UserService} from "../../shared/api";
+import { LoadingComponent } from "src/app/shared/loading/loading.component";
+import { TpHeaderComponent } from "src/app/shared/tp-header/tp-header.component";
+
 import { UsersComponent } from "./users.component";
 
 describe("UsersComponent", () => {
@@ -29,14 +30,6 @@ describe("UsersComponent", () => {
 
 	beforeEach(waitForAsync(() => {
 		// mock the API
-		const mockAPIService = jasmine.createSpyObj(["getUsers", "getRoles", "getCurrentUser"]);
-		mockAPIService.getUsers.and.returnValue(new Promise(resolve => resolve([])));
-		mockAPIService.getRoles.and.returnValue(new Promise(resolve => resolve([])));
-		mockAPIService.getCurrentUser.and.returnValue(new Promise(resolve => resolve({
-			id: 0,
-			newUser: false,
-			username: "test"
-		} as User)));
 		const mockCurrentUserService = jasmine.createSpyObj(["updateCurrentUser", "login", "logout"]);
 		mockCurrentUserService.updateCurrentUser.and.returnValue(new Promise(r => r(false)));
 
@@ -47,13 +40,13 @@ describe("UsersComponent", () => {
 				TpHeaderComponent,
 			],
 			imports: [
+				APITestingModule,
 				FormsModule,
 				HttpClientModule,
 				ReactiveFormsModule,
 				RouterTestingModule
 			],
 			providers: [
-				{ provide: UserService, useValue: mockAPIService },
 				{ provide: CurrentUserService, useValue: mockCurrentUserService }
 			]
 		});
@@ -68,6 +61,80 @@ describe("UsersComponent", () => {
 
 	it("should exist", () => {
 		expect(component).toBeTruthy();
+	});
+
+	it("can tell if a user has a location", () => {
+		const u: User = {
+			city: "Townsville",
+			country: "Countryland",
+			id: -1,
+			newUser: false,
+			postalCode: "00000",
+			stateOrProvince: "Provincia",
+			username: "test"
+		};
+		expect(component.userHasLocation(u)).toBeTrue();
+		u.city = null;
+		expect(component.userHasLocation(u)).toBeTrue();
+		u.stateOrProvince = null;
+		expect(component.userHasLocation(u)).toBeTrue();
+		u.country = null;
+		expect(component.userHasLocation(u)).toBeTrue();
+		u.postalCode = null;
+		expect(component.userHasLocation(u)).toBeFalse();
+		u.country = "Countryland";
+		expect(component.userHasLocation(u)).toBeTrue();
+		u.country = null;
+		u.stateOrProvince = "Provincia";
+		expect(component.userHasLocation(u)).toBeTrue();
+		u.stateOrProvince = null;
+		u.city = "Townsville";
+		expect(component.userHasLocation(u)).toBeTrue();
+	});
+
+	it("builds user location strings", () => {
+		const u: User = {
+			city: "Townsville",
+			country: "Countryland",
+			id: -1,
+			newUser: false,
+			postalCode: "00000",
+			stateOrProvince: "Provincia",
+			username: "test"
+		};
+		expect(component.userLocationString(u)).toBe("Townsville, Provincia, Countryland, 00000");
+		u.city = null;
+		expect(component.userLocationString(u)).toBe("Provincia, Countryland, 00000");
+		u.stateOrProvince = null;
+		expect(component.userLocationString(u)).toBe("Countryland, 00000");
+		u.country = null;
+		expect(component.userLocationString(u)).toBe("00000");
+		u.postalCode = null;
+		expect(component.userLocationString(u)).toBeNull();
+		u.country = "Countryland";
+		expect(component.userLocationString(u)).toBe("Countryland");
+		u.country = null;
+		u.stateOrProvince = "Provincia";
+		expect(component.userLocationString(u)).toBe("Provincia");
+		u.stateOrProvince = null;
+		u.city = "Townsville";
+		expect(component.userLocationString(u)).toBe("Townsville");
+	});
+
+	it("searches fuzz-ily", ()=>{
+		const u = {
+			id: -1,
+			newUser: false,
+			username: "test"
+		};
+		expect(component.fuzzControl.value).toBe("");
+		expect(component.fuzzy(u)).toBeTrue();
+		component.fuzzControl.setValue(`${u.username}z`);
+		expect(component.fuzzy(u)).toBeFalse();
+		component.fuzzControl.setValue(u.username);
+		expect(component.fuzzy(u)).toBeTrue();
+		component.fuzzControl.setValue(`${u.username[0]}${u.username.slice(-1)[0]}`);
+		expect(component.fuzzy(u)).toBeTrue();
 	});
 
 	afterAll(() => {
