@@ -11,8 +11,8 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
-import { MatSnackBar } from "@angular/material/snack-bar";
 import { Component, OnDestroy } from "@angular/core";
+import { MatSnackBar } from "@angular/material/snack-bar";
 import { Subscription } from "rxjs";
 
 import { AlertService } from "./alert.service";
@@ -58,9 +58,6 @@ export class AlertComponent implements OnDestroy {
 							break;
 						case "error":
 							console.error("alert: ", a.text);
-							break;
-						default:
-							console.log("unknown alert: ", a.text);
 							break;
 					}
 					this.snackBar.open(a.text, "dismiss", {duration: this.duration, verticalPosition: "top"});
