@@ -11,36 +11,26 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
+import { CommonModule } from "@angular/common";
+import { HTTP_INTERCEPTORS } from "@angular/common/http";
 import { NgModule } from "@angular/core";
-import {HTTP_INTERCEPTORS} from "@angular/common/http";
-import {CommonModule} from "@angular/common";
-import {RouterModule} from "@angular/router";
-import {AppUIModule} from "../app.ui.module";
-import {AlertComponent} from "./alert/alert.component";
-import {ErrorInterceptor} from "./interceptor/error.interceptor";
-import {AlertInterceptor} from "./interceptor/alerts.interceptor";
-import {LoadingComponent} from "./loading/loading.component";
-import {TpHeaderComponent} from "./tp-header/tp-header.component";
-import {GenericTableComponent} from "./generic-table/generic-table.component";
-import {LinechartDirective} from "./charts/linechart.directive";
-import {AlertService} from "./alert/alert.service";
-import {
-	CacheGroupService,
-	CDNService,
-	DeliveryServiceService,
-	InvalidationJobService,
-	ProfileService,
-	ServerService, TypeService, UserService
-} from "./api";
-import {PhysicalLocationService} from "./api/PhysicalLocationService";
-import {CurrentUserService} from "./currentUser/current-user.service";
-import {CustomvalidityDirective} from "./validation/customvalidity.directive";
-import {OpenableDirective} from "./openable/openable.directive";
-import {SSHCellRendererComponent} from "./table-components/ssh-cell-renderer/ssh-cell-renderer.component";
-import {UpdateCellRendererComponent} from "./table-components/update-cell-renderer/update-cell-renderer.component";
-import {BooleanFilterComponent} from "./table-components/boolean-filter/boolean-filter.component";
+import { RouterModule } from "@angular/router";
 
+import { AppUIModule } from "src/app/app.ui.module";
 
+import { AlertComponent } from "./alert/alert.component";
+import { AlertService } from "./alert/alert.service";
+import { LinechartDirective } from "./charts/linechart.directive";
+import { CurrentUserService } from "./currentUser/current-user.service";
+import { GenericTableComponent } from "./generic-table/generic-table.component";
+import { AlertInterceptor } from "./interceptor/alerts.interceptor";
+import { ErrorInterceptor } from "./interceptor/error.interceptor";
+import { LoadingComponent } from "./loading/loading.component";
+import { BooleanFilterComponent } from "./table-components/boolean-filter/boolean-filter.component";
+import { SSHCellRendererComponent } from "./table-components/ssh-cell-renderer/ssh-cell-renderer.component";
+import { UpdateCellRendererComponent } from "./table-components/update-cell-renderer/update-cell-renderer.component";
+import { TpHeaderComponent } from "./tp-header/tp-header.component";
+import { CustomvalidityDirective } from "./validation/customvalidity.directive";
 
 /**
  * SharedModule contains common code that modules can import independently.
@@ -53,13 +43,9 @@ import {BooleanFilterComponent} from "./table-components/boolean-filter/boolean-
 		GenericTableComponent,
 		BooleanFilterComponent,
 		UpdateCellRendererComponent,
-
 		CustomvalidityDirective,
 		LinechartDirective,
-		OpenableDirective
-	],
-	entryComponents: [
-		SSHCellRendererComponent
+		SSHCellRendererComponent,
 	],
 	exports: [
 		AlertComponent,
@@ -68,10 +54,8 @@ import {BooleanFilterComponent} from "./table-components/boolean-filter/boolean-
 		GenericTableComponent,
 		BooleanFilterComponent,
 		UpdateCellRendererComponent,
-
 		CustomvalidityDirective,
 		LinechartDirective,
-		OpenableDirective
 	],
 	imports: [
 		AppUIModule,
@@ -79,19 +63,10 @@ import {BooleanFilterComponent} from "./table-components/boolean-filter/boolean-
 		RouterModule
 	],
 	providers: [
-		{multi: true, provide: HTTP_INTERCEPTORS, useClass: ErrorInterceptor},
-		{multi: true, provide: HTTP_INTERCEPTORS, useClass: AlertInterceptor},
+		{ multi: true, provide: HTTP_INTERCEPTORS, useClass: ErrorInterceptor },
+		{ multi: true, provide: HTTP_INTERCEPTORS, useClass: AlertInterceptor },
 		AlertService,
-		CacheGroupService,
-		CDNService,
 		CurrentUserService,
-		DeliveryServiceService,
-		InvalidationJobService,
-		PhysicalLocationService,
-		ProfileService,
-		ServerService,
-		TypeService,
-		UserService
-	],
+	]
 })
 export class SharedModule { }
