@@ -12,10 +12,12 @@
 * limitations under the License.
 */
 import { HttpClientModule } from "@angular/common/http";
-import { waitForAsync, ComponentFixture, TestBed } from "@angular/core/testing";
+import { waitForAsync, type ComponentFixture, TestBed } from "@angular/core/testing";
 import { RouterTestingModule } from "@angular/router/testing";
+import type { ICellRendererParams } from "ag-grid-community";
 
 import { CurrentUserService } from "src/app/shared/currentUser/current-user.service";
+
 import { SSHCellRendererComponent } from "./ssh-cell-renderer.component";
 
 describe("SshCellRendererComponent", () => {
@@ -23,14 +25,17 @@ describe("SshCellRendererComponent", () => {
 	let fixture: ComponentFixture<SSHCellRendererComponent>;
 
 	beforeEach(waitForAsync(() => {
-		const mockCurrentUserService = jasmine.createSpyObj(["updateCurrentUser", "login", "logout"]);
-		mockCurrentUserService.updateCurrentUser.and.returnValue(new Promise(r => r(false)));
+		const mockCurrentUserService = jasmine.createSpyObj(
+			["updateCurrentUser", "login", "logout"],
+			{currentUser: {username: "test-admin"}}
+		);
+		mockCurrentUserService.updateCurrentUser.and.returnValue(new Promise(r => r(true)));
+
 		TestBed.configureTestingModule({
 			declarations: [ SSHCellRendererComponent ],
 			imports: [HttpClientModule, RouterTestingModule],
 			providers: [ { provide: CurrentUserService, useValue: mockCurrentUserService} ]
-		})
-			.compileComponents();
+		}).compileComponents();
 	}));
 
 	beforeEach(() => {
@@ -41,5 +46,23 @@ describe("SshCellRendererComponent", () => {
 
 	it("should create", () => {
 		expect(component).toBeTruthy();
+	});
+
+	it("initializes", () => {
+		component.agInit({value: "192.0.2.1"} as ICellRendererParams);
+		expect(component.value).toBe("192.0.2.1");
+
+		component.agInit({value: "192.0.2.2"} as ICellRendererParams);
+		expect(component.value).toBe("192.0.2.2");
+	});
+
+	it("refreshes", () => {
+		let ret = component.refresh({value: "192.0.2.1"} as ICellRendererParams);
+		expect(ret).toBeTrue();
+		expect(component.value).toBe("192.0.2.1");
+
+		ret = component.refresh({value: "192.0.2.2"} as ICellRendererParams);
+		expect(ret).toBeTrue();
+		expect(component.value).toBe("192.0.2.2");
 	});
 });
